requirements.txt
langchain>=0.2.14
langchain-openai>=0.2.2
openai>=1.51.0
tiktoken>=0.7.0
pydantic>=2.8.2
pandas>=2.2.2
tqdm>=4.66.4
python-dotenv>=1.0.1
kaggle>=1.6.14
rich>=13.7.1
